Naming and convention modifications
